Improve UI, contain business logic in class
